requirements.txt
anthropic>=0.34.0
notion-client>=2.2.1
python-dotenv>=1.0.1
pydantic>=2.8.2
backoff>=2.2.1
orjson>=3.10.7
python-dateutil>=2.9.0.post0
httpx==0.27.2